New tab, routing and navigation (#78)
Created a new tab in desktop mode called "Tip", where the Tip
component will be displayed in a modal, just like the Payment
modal. Also, created routing so that it is possible to navigate to
/address/:id/tip and make this modal open, both from desktop and from
mobile. Finally, added a "lastTab" state entry so the dapp knows which
tab to open once it is closed by a user. Created a dummy QR code
inside the Tip modal.